moving pnp helpers to sample
