+++ device/iot-device-samples/pnp-device-sample/temperature-controller-device-sample/src/main/java/samples/com/microsoft/azure/sdk/iot/device/TemperatureController.java
@@ -7,7 +7,6 @@ import com.google.gson.Gson;
 import com.microsoft.azure.sdk.iot.deps.twin.TwinCollection;
 import com.microsoft.azure.sdk.iot.device.*;
 import com.microsoft.azure.sdk.iot.device.DeviceTwin.*;
-import com.microsoft.azure.sdk.iot.device.plugandplay.PnpHelper;
 import com.microsoft.azure.sdk.iot.provisioning.device.*;
 import com.microsoft.azure.sdk.iot.provisioning.device.internal.exceptions.ProvisioningDeviceClientException;
 import com.microsoft.azure.sdk.iot.provisioning.security.SecurityProviderSymmetricKey;
@@ -371,7 +370,7 @@ public class TemperatureController {
                 double targetTemperature = (double) ((TwinCollection) property.getValue()).get(propertyName);
                 log.debug("Property: Received - component=\"{}\", {\"{}\": {}°C}.", componentName, propertyName, targetTemperature);
 
-                Set<Property> pendingPropertyPatch = PnpHelper.createPropertyEmbeddedValuePatch(
+                Set<Property> pendingPropertyPatch = PnpHelpers.createPropertyEmbeddedValuePatch(
                         propertyName,
                         targetTemperature,
                         componentName,
@@ -388,7 +387,7 @@ public class TemperatureController {
                     Thread.sleep(5 * 1000);
                 }
 
-                Set<Property> completedPropertyPatch = PnpHelper.createPropertyEmbeddedValuePatch(
+                Set<Property> completedPropertyPatch = PnpHelpers.createPropertyEmbeddedValuePatch(
                         propertyName,
                         temperature.get(componentName),
                         componentName,
@@ -409,14 +408,14 @@ public class TemperatureController {
         // TODO: Environment.WorkingSet equivalent in Java
         double workingSet = 1024;
 
-        Message message = PnpHelper.createIotHubMessageUtf8(telemetryName, workingSet);
+        Message message = PnpHelpers.createIotHubMessageUtf8(telemetryName, workingSet);
         deviceClient.sendEventAsync(message, new MessageIotHubEventCallback(), message);
         log.debug("Telemetry: Sent - {\"{}\": {}KiB }", telemetryName, workingSet);
     }
 
     private static void sendDeviceSerialNumber() throws IOException {
         String propertyName = "serialNumber";
-        Set<Property> propertyPatch = PnpHelper.createPropertyPatch(propertyName, SERIAL_NO);
+        Set<Property> propertyPatch = PnpHelpers.createPropertyPatch(propertyName, SERIAL_NO);
 
         deviceClient.sendReportedProperties(propertyPatch);
         log.debug("Property: Update - {\"{}\": {}} is {}", propertyName, SERIAL_NO, StatusCode.COMPLETED);
@@ -436,7 +435,7 @@ public class TemperatureController {
         String telemetryName = "temperature";
         double currentTemperature = temperature.get(componentName);
 
-        Message message = PnpHelper.createIotHubMessageUtf8(telemetryName, currentTemperature, componentName);
+        Message message = PnpHelpers.createIotHubMessageUtf8(telemetryName, currentTemperature, componentName);
         deviceClient.sendEventAsync(message, new MessageIotHubEventCallback(), message);
         log.debug("Telemetry: Sent - {\"{}\": {}°C} with message Id {}.", telemetryName, currentTemperature, message.getMessageId());
 
@@ -455,7 +454,7 @@ public class TemperatureController {
         String propertyName = "maxTempSinceLastReboot";
         double maxTemp = maxTemperature.get(componentName);
 
-        Set<Property> reportedProperty = PnpHelper.createPropertyPatch(propertyName, maxTemp, componentName);
+        Set<Property> reportedProperty = PnpHelpers.createPropertyPatch(propertyName, maxTemp, componentName);
         deviceClient.sendReportedProperties(reportedProperty);
         log.debug("Property: Update - {\"{}\": {}°C} is {}.", propertyName, maxTemp, StatusCode.COMPLETED);
     }
