+++ device/pom.xml
@@ -20,6 +20,7 @@
     <modules>
         <module>iot-device-client</module>
         <module>iot-device-samples</module>
+        <module>pnp-helper</module>
     </modules>
     <build>
         <plugins>
